reflection_requirements.txt
# Universal Self-Reflection System Requirements
pyyaml>=6.0
numpy>=1.24.0
